Revert "fix: batched merkle tree zeroing (#2045)"
This reverts commit 435e679448d76c784dd15425e03d8d900e0b95d7.
