--- program-tests/batched-merkle-tree-test/tests/e2e.rs
@@ -1,2 +1,0 @@
-#[path = "e2e_tests/mod.rs"]
-mod e2e_tests;
