--- program-tests/batched-merkle-tree-test/tests/e2e_tests/mod.rs
@@ -1,4 +1,0 @@
-pub mod address;
-pub mod shared;
-pub mod simulate_txs;
-pub mod state;
